\name{plot.mBCV_obj}
\alias{plot.mBCV_obj}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Plot Kernel Density Result from \code{mBCV_obj}
}
\description{
draw a histogram and density curve of the results.
}
\usage{
\method{plot}{mBCV_obj}(x, main=NULL, xlab="", \dots)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{ \code{mBCV_obj} object.}
  \item{main}{title of plot.}
  \item{xlab}{title for the x axis.}
  \item{\dots}{arguments to be paseed to methods.}
}
\value{
No return value. Called for its side effects (generates a plot).
}

