\name{Rcpp_cKDE-class}
\docType{class}
\alias{Rcpp_cKDE-class}
\title{Internal S4 Class for Bandwidth Estimation}
\description{
An internal S4 class used in the implementation of kernel density bandwidth estimation using MCMC samples. This class is not intended to be used directly by package users.
}
\section{Objects from the Class}{
Objects of this class are created internally by C++ functions exposed via Rcpp. End users should not create or manipulate instances of this class.
}
\section{Slots}{
This class does not expose any user-accessible slots.
}
\keyword{internal}